scan newly protected segments while their generation is Gfrom, rather than first promoting them to Gprot.
the bug this fixes was observed by doing the following
interactively (it must be interactive because you need
to compile new code each time):
	gc();
	gc();
	gc();
	gc();
	gc();
	gc();
	gc();
	gc(1,1);
	gc(1,1);

we think the problem was that on the first gc(),
litdom->ns was in the same segment as the segment
containing the newly protected litdom.
thus litdom->ns was not copied because of
the (s->gen != Gfrom) check in copy().
symptoms were that on every compile of the gc calls,
konimm cvals had older, already forwarded type
fields.  these types came from litdom->ns (and
probably were in the protected segment as well,
but this was not checked).