Fix boundary condition in big segment scanning
tl;dr:

Slightly different invariants for large object storage ("big segments")
allow the garbage collector to erroneously visit arbitrary segments
during scanning, potentially leading to livelock (see bug).

Gory details:

Big segments (those used for single-object allocations in excess of a
certain limit) have different invariants than normal segments. In normal
segments, the allocation pointer can never exceed the closed interval
defined by segment base + segment size, because a pointer slot is
reserved at the end of the segment (this slot is not used for anything;
it was added in 55dc73c2 when the scan pointer was moved to per-metatype
specifically to address this problem in normal segments. Big segments
didn't exist at that time).

In any case, the "allocation pointer" for big segments can be equal to
segement base + segment size (scare quotes because it is never actually
used for subsequent allocations, as bigsegs are single-object). This
triggers faulty logic in the scan phase, where we bump the generational
collector's ->scan pointer object + sz.

If it so happens that the last segment in the segment list is a bigseg
with an object that is 0 - 4 bytes smaller than the segment size, the
scan pointer can be left pointing to the next segment _in the system
virtual address space_. A subsequent scan pass of this generation will
then visit the relevant segment, triggering one of three failure
conditions:

  1. bugcheck, if you are lucky and the segment is dead
  2. segfault, if you deference bad memory scanning garbage
  3. livelock, if the segment preceeds you in the same generation's
  segment list

You can easily reproduce this issue with the following statement:

   foo = mkstr((1<<20) - 24);

Where 2**20 is the segment size and 24 is magic internal structure
overhead for strings. This statement leads to an allocation of an object
of exactly the segement size. On a fresh l1 instance this will nicely
bugcheck.